requirements.txt
certifi==2020.6.20
cloudpickle==1.6.0
cycler==0.10.0
decorator==4.4.2
future==0.18.2
imageio==2.9.0
install==1.3.3
joblib==0.16.0
kiwisolver==1.2.0
matplotlib==3.3.2
networkx==2.5
numpy==1.19.2
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
submitit==1.1.1
threadpoolctl==2.1.0
tifffile==2020.9.22
torch==1.6.0
torchvision==0.7.0
typing-extensions==3.7.4.3
